Rename send lifecycle hook per feedback
